pyproject.toml
[tool.poetry]
name = "farpostbooks_backend"
version = "0.1.0"
description = ""
authors = [

]
maintainers = [

]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.92.0"
python-jose="^3.3.0"
uvicorn = { version = "^0.20.0", extras = ["standard"] }
pydantic = {version = "^1.10.5", extras = ["dotenv"]}
yarl = "^1.8.2"
ujson = "^5.7.0"
tortoise-orm = "^0.19.2"
aerich = "^0.7.1"
asyncpg = "^0.27.0"
httptools = "^0.5.0"
sentry-sdk = "^1.14.0"
loguru = "^0.6.0"
jose = "^1.0.0"
aiofiles = "^23.1.0"
types-aiofiles = "^22.1.0.8"

[tool.poetry.dev-dependencies]
pytest = "^7.2.1"
flake8 = "~4.0.1"
mypy = "^1.0.0"
isort = "^5.11.4"
pre-commit = "^3.0.1"
wemake-python-styleguide = "^0.17.0"
black = "^23.1.0"
autoflake = "^1.6.1"
pytest-cov = "^4.0.0"
anyio = "^3.6.2"
pytest-env = "^0.8.1"
asynctest = "^0.13.0"
nest-asyncio = "^1.5.6"
httpx = "^0.23.3"
Faker="^17.0.0"

[tool.isort]
profile = "black"
multi_line_output = 3
src_paths = ["farpostbooks_backend",]

[tool.mypy]
strict = true
ignore_missing_imports = true
allow_subclassing_any = true
allow_untyped_calls = true
pretty = true
show_error_codes = true
implicit_reexport = true
allow_untyped_decorators = true
warn_unused_ignores = false
warn_return_any = false
namespace_packages = true

[tool.pytest.ini_options]
filterwarnings = [
    "error",
    "ignore::DeprecationWarning",
    "ignore:.*unclosed.*:ResourceWarning",
]
env = [
    "FARPOSTBOOKS_BACKEND_DB_BASE=farpostbooks_backend_test",
    "FARPOSTBOOKS_BACKEND_SENTRY_DSN=",
]
[tool.aerich]
tortoise_orm = "farpostbooks_backend.db.config.TORTOISE_CONFIG"
location = "./farpostbooks_backend/db/migrations"
src_folder = "./farpostbooks_backend"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
